New art and sounds
- Added new art and sounds from opengameart.org
- Added setVolume method to Ball class
- Added setMusic, setVolum, and play methods to screen class.
- Made paddles and ball smaller
- Increased the speed of the right paddle, as it could not keep up with
  the smaller ball.

+++ Pong/Paddle.cpp
@@ -18,7 +18,7 @@ namespace sf {
 		return true;
 	}
 
-	void Paddle::setVelocity(int velocity) {
+	void Paddle::setVelocity(double velocity) {
 		this->velocity = velocity;
 	}
